Removed font files import from mds
There is an inconsistence between libraries as we are importing the default fonts in the applications by using GlobalStyles. The disadvantage of this is that there is a flickering between renders in the initial load until the fonts get loaded.

This PR removes de default font imports by levering this to the application where you are going to use mds.

Signed-off-by: Benjamin Perez <benjamin@bexsoft.net>